requirements.txt
numpy==1.19.5
streamlit
matplotlib==3.1.0
pandas==1.0.1
Pillow==8.1.0
tensorflow-cpu==2.4.1
google-cloud-aiplatform